Using J-Unit in IntelliJ
With Intellij, J-unit 5 is available withthe newest Version. I Used JUnit 5 to conduct my Tests, However we imported Junit 4.2 from our Pom.xml.
To Create a Test in IntelliJ, You must click on the name of the class You are looking to test. Once it is Highlighted, Press Alt+Enter and you are given the Option to create a test. From There you create a test with the testing material you want and for us it was Junit. 
From There You import the Junit material and go ahead. 
Assertions in Junit help to find problems when testing the code.